requirements.txt
Pillow==6.1.0
# csv
pyyaml